Separate FromDBStore and CommonStore
Previously, Store (renamed to FromDBStore) and CommonStore
interfaces shared the same file (storage.go). In order to improve
readability, we have moved CommonStore into its own file. Each
...Store interface clearly shows the databases it connects to.
In the future, we can implement a ...Store interface for toDB
and if we need to connect only to information_schema, we could
implment an IschemaStore interface.
